refactor: SignUpForm 컴포넌트 내에서 처리
